+++ lib/lexer/position.dart
@@ -1,0 +1,6 @@
+class Cursor {
+  int line;
+  int position;
+
+  Cursor({this.line = 1, this.position = 0});
+}