+++ CHANGELOG.md
@@ -8,4 +8,12 @@
 
 ## 1.1.0
 
-- Bug fixes and final release
+- Bug fixes 
+
+## 1.1.1
+
+- Bug fixes: Problems with path for finding `todo.txt`
+
+## 1.1.2
+
+- Refactor: Lexer and parser loops